refactor: move write_tacs to regional_tac_extraction
