+++ petpal/cli/cli_preproc.py
@@ -87,6 +87,8 @@ See Also:
 import argparse
 import ants
 
+import petpal.preproc.regional_tac_extraction
+
 from ..utils import useful_functions
 from ..preproc import image_operations_4d, motion_corr, register
 
@@ -352,7 +354,7 @@ def main():
                               half_life=args.half_life)
 
     if command=='write_tacs':
-        image_operations_4d.write_tacs(input_image_path=args.input_img,
+        petpal.preproc.regional_tac_extraction.write_tacs(input_image_path=args.input_img,
                                        out_tac_dir=args.out_tac_dir,
                                        segmentation_image_path=args.segmentation,
                                        label_map_path=args.label_map_path,
